Refactor package structure - Move animal domain to a 'domain' package - Move security components to 'security' package - Extract security concerns from domain, move to their own Controllers and tests
