+++ backend/src/main/java/io/spring/cloud/samples/animalrescue/backend/domain/AdoptionRequest.java
@@ -1,4 +1,4 @@
-package io.spring.cloud.samples.animalrescue.backend;
+package io.spring.cloud.samples.animalrescue.backend.domain;
 
 import org.springframework.data.annotation.Id;
