+++ backend/src/main/java/io/spring/cloud/samples/animalrescue/backend/security/KubernetesSecurityConfiguration.java
@@ -1,4 +1,4 @@
-package io.spring.cloud.samples.animalrescue.backend;
+package io.spring.cloud.samples.animalrescue.backend.security;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
