+++ backend/src/main/java/io/spring/cloud/samples/animalrescue/backend/security/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package io.spring.cloud.samples.animalrescue.backend;
+package io.spring.cloud.samples.animalrescue.backend.security;
 
 import java.net.URI;
