+++ backend/src/main/java/io/spring/cloud/samples/animalrescue/backend/security/SecurityController.java
@@ -1,0 +1,41 @@
+package io.spring.cloud.samples.animalrescue.backend.security;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.security.Principal;
+
+@RestController
+public class SecurityController {
+
+	private static final Logger LOGGER = LoggerFactory.getLogger(SecurityController.class);
+
+	public SecurityController() {
+	}
+
+	@GetMapping("/whoami")
+	public String whoami(Principal principal) {
+		if (principal == null) {
+			return "";
+		}
+		return principal.getName();
+	}
+
+	@ExceptionHandler({AccessDeniedException.class})
+	public ResponseEntity<String> handleAccessDeniedException(Exception e) {
+		return new ResponseEntity<>(e.getMessage(), new HttpHeaders(), HttpStatus.FORBIDDEN);
+	}
+
+	@ExceptionHandler({IllegalArgumentException.class})
+	public ResponseEntity<String> handleIllegalArgumentException(Exception e) {
+		return new ResponseEntity<>(e.getMessage(), new HttpHeaders(), HttpStatus.BAD_REQUEST);
+	}
+
+}
